Added authentication to todos

+++ fastapi/TodoApp/routers/todos.py
@@ -2,11 +2,10 @@ from typing import Annotated
 from pydantic import BaseModel, Field
 from fastapi import FastAPI, Depends, HTTPException, Path, APIRouter
 from sqlalchemy.orm import Session
-import models
 from models import *
-from database import engine, SessionLocal
+from database import SessionLocal
 from starlette import status
-from routers import auth
+from .auth import get_current_user
 
 router = APIRouter(prefix="/todos", tags=["todos"])
 
@@ -20,7 +19,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
-
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -30,29 +29,43 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user:user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+    return db.query(Todos).filter(Todos.owner_id == user['id']).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user:user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user['id']).first())
+
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found.")
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user:user_dependency, db: db_dependency, todo_request: TodoRequest):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+
+    todo_model = Todos(**todo_request.model_dump(), owner_id = user['id'])
 
     db.add(todo_model)
     db.commit()
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def update_todo(db: db_dependency, todo_id: int, todo_request: TodoRequest):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_id: int, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+
+    todo_model = db.query(Todos).filter(Todos.owner_id == user['id']).filter(Todos.id == todo_id).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
@@ -64,12 +77,14 @@ async def update_todo(db: db_dependency, todo_id: int, todo_request: TodoRequest
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+
+    todo_model = db.query(Todos).filter(Todos.owner_id == user['id']).filter(Todos.id == todo_id).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
-
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user['id']).delete()
     db.commit()
